Rejig code for the machine ID extraction
