+++ source/server/remote_dispatch_common.c
@@ -3,11 +3,6 @@
 #ifdef _WIN32
 // see ReflectiveLoader.c...
 extern HINSTANCE hAppInstance;
-#else
-#include <sys/types.h>
-#include <dirent.h>
-#include <sys/utsname.h>
-#define MAX_PATH 256
 #endif
 
 PLIST gExtensionList = NULL;
@@ -52,71 +47,6 @@ BOOL ext_cmd_callback(LPVOID pState, LPVOID pData)
 	return FALSE;
 }
 
-DWORD request_core_machine_id(Remote* pRemote, Packet* pPacket)
-{
-	DWORD res = ERROR_SUCCESS;
-	Packet* pResponse = packet_create_response(pPacket);
-
-	if (pResponse)
-	{
-#ifdef _WIN32
-		wchar_t buffer[MAX_PATH];
-		if (GetSystemDirectory(buffer, MAX_PATH) != 0)
-		{
-			wchar_t computerName[MAX_PATH];
-			DWORD computerNameSize = MAX_PATH;
-			DWORD serialNumber;
-			wchar_t* backslash = wcschr(buffer, L'\\');
-			*(backslash + 1) = L'\0';
-
-			GetVolumeInformation(buffer, NULL, 0, &serialNumber, NULL, 0, NULL, 0);
-
-			GetComputerName(computerName, &computerNameSize);
-
-			_snwprintf_s(buffer, MAX_PATH, MAX_PATH - 1, L"%04x-%04x:%s", HIWORD(serialNumber), LOWORD(serialNumber), computerName);
-			packet_add_tlv_wstring(pResponse, TLV_TYPE_MACHINE_ID, buffer);
-		}
-#else
-		char buffer[MAX_PATH];
-    struct dirent *data;
-    struct utsname utsbuf;
-    char* folder = "/dev/disk/by-id/";
-    DIR *ctx = opendir(folder);
-
-    if (uname(&utsbuf) == -1)
-    {
-      res = GetLastError();
-      goto out;
-    }
-
-    if (ctx == NULL)
-    {
-      res = GetLastError();
-      goto out;
-    }
-
-    while (data = readdir(ctx))
-    {
-      // TODO: make sure that looking for drives prefixed with "ata" is a good
-      // idea. We might need to search for a bunch of prefixes.
-      if (strncmp(data->d_name, "ata-", 4) == 0)
-      {
-        snprintf(buffer, MAX_PATH - 1, "%s:%s", data->d_name + 4, utsbuf.nodename);
-        packet_add_tlv_string(pResponse, TLV_TYPE_MACHINE_ID, buffer);
-        break;
-      }
-    }
-    closedir(ctx);
-
-out:
-
-#endif
-
-		packet_transmit_response(res, pRemote, pResponse);
-	}
-
-	return ERROR_SUCCESS;
-}
 
 DWORD request_core_enumextcmd(Remote* pRemote, Packet* pPacket)
 {
